Rename files to use underscores instead of hyphens
