+++ components/brave_wallet_ui/components/desktop/hardware-wallet-connect/hardware_wallet_connect.types.ts
