+++ components/brave_wallet_ui/components/desktop/popup-modals/add-account-modal/hardware-button/hardware_button.style.ts
