+++ components/brave_wallet_ui/components/desktop/popup-modals/add-account-modal/hardware-button/hardware_button.tsx
@@ -12,7 +12,7 @@ import {
   Title,
   ArrowIcon,
   Description
-} from './hardware-button.style'
+} from './hardware_button.style'
 
 interface Props {
   title: string
