+++ components/brave_wallet_ui/components/shared/tooltip/password-strength-tooltip.stories.tsx
@@ -19,7 +19,7 @@ export const _PasswordStrengthTooltip = () => {
       />
       <Button
         onClick={() => {
-          setIsVisible(!isVisible)
+          setIsVisible((prev) => !prev)
         }}
       >
         Hover
