+++ components/brave_wallet_ui/page/screens/onboarding/components/auto_lock_settings/auto_lock_settings.style.ts
