+++ components/brave_wallet_ui/page/screens/onboarding/components/auto_lock_settings/auto_lock_settings.tsx
@@ -16,7 +16,7 @@ import {
   LockIconContainer,
   SettingDescription,
   OptionLabel
-} from './auto-lock-settings.style'
+} from './auto_lock_settings.style'
 import { getLocale } from '../../../../../../common/locale'
 
 interface Props {
