+++ components/brave_wallet_ui/page/screens/onboarding/components/onboarding-import-wallet-type/import_wallet_type.tsx
@@ -11,7 +11,7 @@ import { WalletRoutes } from '../../../../../constants/types'
 import { getLocale } from '../../../../../../common/locale'
 
 // components
-import { OnboardingContentLayout } from '../onboarding-content-layout/content-layout'
+import { OnboardingContentLayout } from '../onboarding_content_layout/content_layout'
 import { ImportTypeAction } from './components/import-type-action'
 
 // styles
