+++ components/brave_wallet_ui/page/screens/onboarding/components/onboarding_content_layout/content_layout.style.ts
