+++ components/brave_wallet_ui/page/screens/onboarding/creating_wallet/onboarding_creating_wallet.style.ts
