+++ components/brave_wallet_ui/page/screens/onboarding/network_selection/onboarding_network_selection.tsx
@@ -50,7 +50,7 @@ import {
 } from '../../../../components/shared/create-network-icon'
 import {
   OnboardingContentLayout //
-} from '../components/onboarding-content-layout/content-layout'
+} from '../components/onboarding_content_layout/content_layout'
 
 // styles
 import {
