+++ components/brave_wallet_ui/page/screens/onboarding/restore-from-recovery-phrase/restore-from-extension.tsx
@@ -41,10 +41,10 @@ import {
 } from '../../../../components/shared/password-input/new-password-input'
 import {
   OnboardingCreatingWallet //
-} from '../creating-wallet/onboarding-creating-wallet'
+} from '../creating_wallet/onboarding_creating_wallet'
 import {
   OnboardingContentLayout //
-} from '../components/onboarding-content-layout/content-layout'
+} from '../components/onboarding_content_layout/content_layout'
 
 // styles
 import { InputLabel } from './restore-from-recovery-phrase.style'
